refactor(design): move title and subtitle into subfolders
