+++ libs/design/src/molecules/hero/hero-subtitle/hero-subtitle.component.spec.ts
