+++ libs/design/src/molecules/hero/hero-title/hero-title.component.ts
