+++ libs/design/src/molecules/hero/public_api.ts
@@ -1,4 +1,4 @@
 export { DaffHeroModule } from './hero.module';
 export { DaffHeroComponent } from './hero.component';
-export { DaffHeroTitleComponent } from './hero-title.component';
-export { DaffHeroSubtitleComponent } from './hero-subtitle.component';
+export { DaffHeroTitleComponent } from './hero-title/hero-title.component';
+export { DaffHeroSubtitleComponent } from './hero-subtitle/hero-subtitle.component';
